+++ src/SolrExpress/Search/Parameter/Validation/SearchParameterIsInvalidException.cs
@@ -4,8 +4,8 @@ namespace SolrExpress.Search.Parameter.Validation
 {
     public class SearchParameterIsInvalidException : Exception
     {
-        public SearchParameterIsInvalidException(string parameterType) :
-            base(string.Format(Resource.SearchParameterIsInvalidException, parameterType))
+        public SearchParameterIsInvalidException(string parameterType, string errorMessage) :
+            base(string.Format(Resource.SearchParameterIsInvalidException, parameterType, errorMessage))
         {
         }
     }